fix: amount to int

+++ internal/usecases/income_repository.go
@@ -9,6 +9,6 @@ import (
 
 // IncomeRepository репозиторий для работы с доходами
 type IncomeRepository interface {
-	AddIncome(ctx context.Context, userID int64, categoryID int, amount float64, description string) error
+	AddIncome(ctx context.Context, userID int64, categoryID int32, amount int64, description string) error
 	GetIncomesForPeriod(ctx context.Context, userID int64, startDate, endDate string) ([]*domain.Income, error)
 }
